+++ src/View/clear/clearLottoInputs.js
