+++ src/View/clear/clearUIElements.js
