+++ src/View/create/createLottoInput.js
