+++ src/View/show/showPurchaseResult.js
